+++ packages/component/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @orderly.network/react
 
+## 1.2.0
+
+### Minor Changes
+
+- 52bebed3: fix bugs;
+
+### Patch Changes
+
+- Updated dependencies [52bebed3]
+  - @orderly.network/trading-view@1.2.0
+  - @orderly.network/hooks@1.2.0
+  - @orderly.network/types@1.2.0
+  - @orderly.network/utils@1.2.0
+  - @orderly.network/core@1.2.0
+  - @orderly.network/net@1.2.0
+
+## 1.2.0-alpha.0
+
+### Minor Changes
+
+- fix bugs;
+
+### Patch Changes
+
+- Updated dependencies
+  - @orderly.network/trading-view@1.2.0-alpha.0
+  - @orderly.network/hooks@1.2.0-alpha.0
+  - @orderly.network/types@1.2.0-alpha.0
+  - @orderly.network/utils@1.2.0-alpha.0
+  - @orderly.network/core@1.2.0-alpha.0
+  - @orderly.network/net@1.2.0-alpha.0
+
+## 0.2.9
+
+### Patch Changes
+
+- fix bug: app crashes when symbolInfo is null;
+- Updated dependencies
+  - @orderly.network/core@0.2.8
+  - @orderly.network/hooks@1.1.9
+  - @orderly.network/net@1.1.8
+  - @orderly.network/trading-view@0.2.9
+  - @orderly.network/types@0.2.8
+  - @orderly.network/utils@0.1.8
+
 ## 0.2.8
 
 ### Patch Changes