+++ packages/onboard/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @orderly.network/onboard
 
+## 1.2.0
+
+### Minor Changes
+
+- 52bebed3: fix bugs;
+
+### Patch Changes
+
+- Updated dependencies [52bebed3]
+  - @orderly.network/hooks@1.2.0
+  - @orderly.network/types@1.2.0
+
+## 1.2.0-alpha.0
+
+### Minor Changes
+
+- fix bugs;
+
+### Patch Changes
+
+- Updated dependencies
+  - @orderly.network/hooks@1.2.0-alpha.0
+  - @orderly.network/types@1.2.0-alpha.0
+
+## 0.1.9
+
+### Patch Changes
+
+- fix bug: app crashes when symbolInfo is null;
+- Updated dependencies
+  - @orderly.network/hooks@1.1.9
+  - @orderly.network/types@0.2.8
+
 ## 0.1.8
 
 ### Patch Changes